Refactor de controladores a servicios y migración completa de DTOs

+++ arfit/backend/src/main/java/com/albertoandraul/arfit/dto/ProgressLogRequestDTO.java
@@ -1,5 +1,6 @@
 package com.albertoandraul.arfit.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record ProgressLogRequestDTO(Long exerciseId, LocalDate date, Double weight, Integer reps, Integer sets) {}
+public record ProgressLogRequestDTO(Long exerciseId, LocalDate date, BigDecimal weight, Integer reps, Integer sets) {}
